requirements.txt
sentence-transformers==2.2.2
pinecone-client
langchain==0.0.225
langchain_community
flask
pypdf
python-dotenv
flask

-e .